# src/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Doormat",
    "description": "AI-first rental finder - autonomous property manager discovery and listing scoring",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/discovery/trigger": {
      "post": {
        "tags": [
          "discovery"
        ],
        "summary": "Trigger Discovery V2",
        "description": "Trigger a discovery run. Creates a persisted run record with logs.",
        "operationId": "trigger_discovery_v2_api_discovery_trigger_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TriggerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoveryRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/discovery/runs": {
      "get": {
        "tags": [
          "discovery"
        ],
        "summary": "List Runs",
        "description": "List recent discovery runs, newest first.",
        "operationId": "list_runs_api_discovery_runs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DiscoveryRunResponse"
                  },
                  "title": "Response List Runs Api Discovery Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/discovery/runs/{run_id}": {
      "get": {
        "tags": [
          "discovery"
        ],
        "summary": "Get Run",
        "description": "Get a single discovery run with all log lines.",
        "operationId": "get_run_api_discovery_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoveryRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/discovery/cities/{city}": {
      "post": {
        "tags": [
          "discovery"
        ],
        "summary": "Trigger Discovery",
        "operationId": "trigger_discovery_api_discovery_cities__city__post",
        "parameters": [
          {
            "name": "city",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "City"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/TriggerRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoveryResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/discovery/cities/{city}/managers": {
      "get": {
        "tags": [
          "discovery"
        ],
        "summary": "List Managers",
        "operationId": "list_managers_api_discovery_cities__city__managers_get",
        "parameters": [
          {
            "name": "city",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "City"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ManagerOut"
                  },
                  "title": "Response List Managers Api Discovery Cities  City  Managers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/discovery/cities/{city}/status": {
      "get": {
        "tags": [
          "discovery"
        ],
        "summary": "City Status",
        "operationId": "city_status_api_discovery_cities__city__status_get",
        "parameters": [
          {
            "name": "city",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "City"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CityStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CityStatus": {
        "properties": {
          "city": {
            "type": "string",
            "title": "City"
          },
          "managers_total": {
            "type": "integer",
            "title": "Managers Total"
          },
          "managers_validated": {
            "type": "integer",
            "title": "Managers Validated"
          },
          "has_been_discovered": {
            "type": "boolean",
            "title": "Has Been Discovered"
          }
        },
        "type": "object",
        "required": [
          "city",
          "managers_total",
          "managers_validated",
          "has_been_discovered"
        ],
        "title": "CityStatus"
      },
...
      "DiscoveryResult": {
        "properties": {
          "city": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "City"
          },
          "candidates_found": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Candidates Found"
          },
          "validated_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Validated Count"
          },
          "cached": {
            "type": "boolean",
            "title": "Cached",
            "default": false
          },
          "cost_usd": {
            "type": "number",
            "minimum": 0.0,
            "title": "Cost Usd"
          },
          "duration_seconds": {
            "type": "number",
            "minimum": 0.0,
            "title": "Duration Seconds"
          }
        },
        "type": "object",
        "required": [
          "city",
          "candidates_found",
          "validated_count",
          "cost_usd",
          "duration_seconds"
        ],
        "title": "DiscoveryResult",
        "description": "Aggregate result of a city discovery run."
      },
      "DiscoveryRunLogResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "sequence": {
            "type": "integer",
            "title": "Sequence"
          },
          "level": {
            "type": "string",
            "title": "Level"
          },
          "component": {
            "type": "string",
            "title": "Component"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "sequence",
          "level",
          "component",
          "message",
          "timestamp"
        ],
        "title": "DiscoveryRunLogResponse"
      },
      "DiscoveryRunResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "preference_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preference Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "managers_found": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Managers Found"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "logs": {
            "items": {
              "$ref": "#/components/schemas/DiscoveryRunLogResponse"
            },
            "type": "array",
            "title": "Logs",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "city",
          "status",
          "started_at"
        ],
        "title": "DiscoveryRunResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ManagerOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "website": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website"
          },
          "listing_page_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Listing Page Url"
          },
          "validated": {
            "type": "boolean",
            "title": "Validated"
          }
        },
        "type": "object",
        "required": [
          "id",
          "city",
          "name",
          "website",
          "listing_page_url",
          "validated"
        ],
        "title": "ManagerOut"
      },
...
      "TriggerRequest": {
        "properties": {
          "city": {
            "type": "string",
            "title": "City"
          },
          "preference_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preference Id"
          }
        },
        "type": "object",
        "required": [
          "city"
        ],
        "title": "TriggerRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}